requirements.txt
python-binance==1.0.19
websockets==12.0
pandas==2.2.2
numpy==1.26.4
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
loguru==0.7.2
PyYAML==6.0.1
pytest==8.3.2
